Add provides to Acme-Foo META.json
t/02_no_index.t was failing on Perls older that 5.18 because (for some
reason) the version for Acme::Foo was parsed as being undef.  By adding
a provides section to the META.json we're no longer relying on
Parse::LocalDistribution to provide the version numbers in this dist and
we're able to get consistent results across Perls 5.12 - 5.20.  Good
news: this fixes master, which has been broken for a long time.  Bad
news: this papers over a version parsing issue in older Perls (at least
on Travis).

I think this is an improvement, but that we should remain aware of the
issue with version as it could likely come up in other places.  I think
it's key to get the test suite working, though.