Revert to plain old version from MetaCPAN rather than version_numified.
The numified version turned out to be too clever and switched some
version numbers from decimal to scientific notation.  Also transforming
the original version led to changes in padding with zeroes, which causes
a lot of diff churn and doesn't add anything helpful to 02packages.  We
already use version.pm to compare versions, so it's probably best to
leave the author's intended version in 02packages where possible.

On the whole, using the numified version seemd to create more problems
than it might solve and complicate the code unnecessarily in trying to
work around these new problems.